refactor: Defer virtual-to-physical translation to a workqueue
The previous implementation performed virtual-to-physical address
translation within the perf event interrupt handlers. This is unsafe
as it involves get_user_pages, which walks the page table & can sleep
while we're in an interrupt context.
Plus, it is conventional to offload this worker to kworkers regardless.

This commit refactors the sampling process to address this issue:

* The sample_t struct is modified to store the mm_struct and virtual
address, rather than the physical page.

* The interrupt handlers now only capture the virtual address and
mm_struct, and queue this information for later processing.

* The virtual-to-physical address translation is now performed in the
action_wq_callback workqueue (sample analysis), which provides a safe context for
this operation.

* virt_to_phy function is updated to use get_user_pages_remote
with proper locking (mmap_read_lock) to safely access another
process's memory map.