fix: Fix interrupt safety using perf events and spinlocks
This commit addresses crashes due to calling `perf_event_read_value` in our timer callback, which can sleep due to grabbing a context lock, making it not hardirq safe.
Move LLC miss tracking and all state tracking to the existing main workqueue.

In addition, we switch from using a spinlock on the sampling state to an
atomic. This was problematic as calling `perf_event_enable` causing a
non-maskable interrupt, which we can't have occur while spinning
(otherwise we also crashed).